Add interface for fluid resizing by resize handle
The idea is to attach the handle to the controller and then calculate
the euclidean distance between the original handle position and the
position of the handle (attached to the controller); some way to
determine the sign of resizing (increase/decrease) has yet to be found.

Some additional changes to driver `multiple-layers.js` particularly in
that `handleToolbarIntersections` now handles engaging the move and
resize interactions. `handleSelectEnd` is still where move and resize
are disengaged.