Use left and right points to expand/compress fluidly
Two points `handleLeftPoint` and `handleRightPoint` are created every time
the resize handle is engaged. Then these points are used to determine
whether to expand or compress.

This way of doing it seems to be pretty buggy—3D euclidean distance between
`currPosition` (position that the controller is pointing at) and the left
and right points are not always reflected properly; you can be pointing to
the right but the distance to the right point is somehow longer than the
distance to the left.

Open to new ways to determine the sign of resizing.

'#' will be ignored, and an empty message aborts the commit.  # # On branch
feat-resizing # Your branch is up to date with 'origin/feat-resizing'.  # #
Changes to be committed: #	modified:
apps/simple-video-layers/multiple-layers.js #	modified:
util/webxr/MediaLayerManager/GlassLayer.js #	modified:
util/webxr/Toolbar.js #